requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
Baker==1.3
Django==4.2.13
django-filter==24.2
django-rest-framework-nested==0.0.1
djangorestframework==3.15.2
exceptiongroup==1.2.1
inflection==0.5.1
iniconfig==2.0.0
model-bakery==1.18.1
numpy==1.24.4
packaging==24.1
pandas==2.0.3
pluggy==1.5.0
pytest==8.2.2
pytest-django==4.8.0
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
sqlparse==0.5.0
tomli==2.0.1
typing_extensions==4.12.2
tzdata==2024.1
